requirements.txt
django
jsonfield
google-api-python-client
pyyaml
pandas
